Ajeitando algumas coisas e resultados de benchmark feitos

+++ src/templates/ep-pp.rs
@@ -181,7 +181,7 @@ fn main() {
         println!("Alguma coisa de errado não etá certo!");
     }
 
-    let mops: f64 = ((1 << (M as i32 + 1)) as f64) / tm / 1000000.0;
+    let mops: f64 = (((1 as i64) << ((M as i64) + 1)) as f64) / tm / 1000000.0;
 
     println!(" EP Benchmark Results:\n");
     println!(" CPU Time = NOT IMPLEMENTED YET");
